Improve code annotations and correct data types
This commit adjusts multiple functions across different modules. It adds or modifies comments for better clarity and corrects some constructors to be explicit. Furthermore, the types in ByteCodeOptimizer_Jumps, ByteCodeDebugger_Print, and ByteCodeRun modules are corrected to prevent potential bugs, and instances of narrowing conversions in the ByteCodeRun module are flagged with an appropriate code review tool note.

+++ src/ByteCodeDebugger_Print.cpp
@@ -476,7 +476,7 @@ BcDbgCommandResult ByteCodeDebugger::cmdMemory(ByteCodeRunContext* context, cons
 
         addrB = addrLine;
         addrB += static_cast<int64_t>(min(count, perLine)) * (fmt.bitCount / 8);
-        count -= static_cast<int64_t>(min(count, perLine));
+        count -= min(count, perLine);
         if (!count)
             break;
     }
